# docs/swagger.yaml
basePath: /
definitions:
...
  models.ShiftType:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      error:
        example: Fehler bei der Verarbeitung
        type: string
      message:
        example: Operation erfolgreich
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Carsten Bröckert
  description: REST-API für die Verwaltung von Schichtplänen
  license:
    name: MIT
  title: Schichtplaner API
  version: "1.0"
paths:
...
  /api/v1/shifttypes:
    get:
      consumes:
      - application/json
      description: Ruft alle Schichttypen mit ihren Beziehungen ab
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ShiftType'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Alle Schichttypen abrufen
      tags:
      - shifttypes
    post:
      consumes:
      - application/json
      description: Erstellt einen neuen Schichttyp mit Validierungen
      parameters:
      - description: Schichttyp-Daten
        in: body
        name: shifttype
        required: true
        schema:
          $ref: '#/definitions/models.ShiftType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichttyp erstellen
      tags:
      - shifttypes
  /api/v1/shifttypes/{id}:
    delete:
      consumes:
      - application/json
      description: Löscht einen Schichttyp
      parameters:
      - description: Schichttyp-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichttyp löschen
      tags:
      - shifttypes
    get:
      consumes:
      - application/json
      description: Ruft einen spezifischen Schichttyp mit Details ab
      parameters:
      - description: Schichttyp-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftType'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Einzelnen Schichttyp abrufen
      tags:
      - shifttypes
    put:
      consumes:
      - application/json
      description: Aktualisiert einen bestehenden Schichttyp
      parameters:
      - description: Schichttyp-ID
        in: path
        name: id
        required: true
        type: integer
      - description: Aktualisierte Schichttyp-Daten
        in: body
        name: shifttype
        required: true
        schema:
          $ref: '#/definitions/models.ShiftType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftType'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichttyp aktualisieren
      tags:
      - shifttypes
  /api/v1/shifttypes/department/{id}:
    get:
      consumes:
      - application/json
      description: Ruft alle Schichttypen einer spezifischen Abteilung ab
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ShiftType'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichttypen einer Abteilung abrufen
      tags:
      - shifttypes
...
swagger: "2.0"